Remove the BaseHandler type
After removing the PrepareXContext methods, there wasn't much left in
this type. Move the installation method to a top-level function and
recommend users embed ClientCreator instances directly.

Users can also chose to create their own base types without conflict
from this library.
